% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-mainWrappers.R
\name{AsSentimentMatrix}
\alias{AsSentimentMatrix}
\title{\code{AsSentimentMatrix}}
\usage{
AsSentimentMatrix(x, remove.stopwords = TRUE, stoplist = ftaStopList,
  operations = c("spelling", "stemming"),
  spelling.dictionary = ftaDictionary, manual.replacements = NULL,
  pos.words = ftaPositiveWords, neg.words = ftaNegativeWords)
}
\arguments{
\item{x}{Either a character vector or a WordBag object, created for example by the function InitializeWordBag.
If \code{x} is a character vector then additional parameters can be supplied for processing the
text before scoring sentiment. For a description of the options see \code{\link{InitializeWordBag}}.}

\item{pos.words}{A character vector containing words that should be scored positively.}

\item{neg.words}{A character vector containing words that should be scored negatively.}
}
\value{
A matrix with three columns, and one row for each of the original text responses in the word bag.
        The first column provides a count of the positive words that have been identified in each
        response, the second column gives a count of the negative words, and the third column gives the
        net sentiment score, which is the difference between the two.
}
\description{
Generate a matrix of sentiment scores from a character vector or wordBag object.
}
\details{
The scoring algorithm attempts to handle negation by looking for words like "not" which occur
         one or two tokens prior to a sentiment word. For example, "not bad" would yield a positive
         score of 1 rather than a negative score of -1 which would be applied for the word "bad".
}
\examples{
AsSentimentMatrix(ftaFavoriteThings)
}

