% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-mainWrappers.R
\name{AsTermMatrix}
\alias{AsTermMatrix}
\title{\code{AsTermMatrix}}
\usage{
AsTermMatrix(x, min.frequency = 5, remove.stopwords = TRUE,
  stoplist = ftaStopList, operations = c("spelling", "stemming"),
  spelling.dictionary = ftaDictionary, manual.replacements = NULL)
}
\arguments{
\item{x}{Either a character vector or a WordBag object, created for example by the function InitializeWordBag.
         If \code{x} is a character vector then additional parameters can be supplied for processing the
         text before scoring sentiment. For a description of the options see \code{\link{InitializeWordBag}}.}
}
\value{
A \code{matrix} with one row for each text response (referred to as a \code{document}) and once column for
         each word or phrase extracted from the text.
}
\description{
Generate a matrix of sentiment scores from a character vector or wordBag object.
}
\examples{
AsTermMatrix(ftaFavoriteThings)
}

