% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{FindSpellingErrors}
\alias{FindSpellingErrors}
\title{\code{FindSpellingErrors}}
\usage{
FindSpellingErrors(x, dictionary = get("ftaDictionary"))
}
\arguments{
\item{x}{A vector of words.}

\item{dictionary}{A vector of words to use as a dictionary. By default this uses an english
language dictionary, \code{ftaDictionary} which comes with this package.}
}
\value{
A binary vector indicating which of the elements of x are misspelled according to the
input dictionary. When a word is not found, the function also looks for other stems of the
word, including \code{word + "s"}, \code{word + "es"}, \code{word + "ed"}, \code{word + "d"},
\code{word + "ise"}, and \code{word + "ize"}.
}
\description{
Identify spelling errors in a vector of strings by comparing with a dictionary.
}
\examples{
FindSpellingErrors(c("hello", "hullo", "hollo", "hollow"))

}
