% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftaDictionary}
\alias{ftaDictionary}
\title{English language dictionary.}
\format{A character vector with 92,439 entries.}
\usage{
ftaDictionary
}
\description{
English language dictionary.
}
\keyword{datasets}
