% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftaFavoriteThings}
\alias{ftaFavoriteThings}
\title{Simple text with spelling mistakes introduced.}
\format{A character vector with 51 entries.}
\usage{
ftaFavoriteThings
}
\description{
Simple text with spelling mistakes introduced.
}
\keyword{datasets}

