% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftaNegativeWords}
\alias{ftaNegativeWords}
\title{List of English words with negative sentiment.}
\format{A character vector with 4,962 entries.}
\usage{
ftaNegativeWords
}
\description{
List of English words with negative sentiment.
}
\keyword{datasets}

