% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftaPositiveWords}
\alias{ftaPositiveWords}
\title{List of English words with positive sentiment.}
\format{A character vector with 2,098 entries.}
\usage{
ftaPositiveWords
}
\description{
List of English words with positive sentiment.
}
\keyword{datasets}

