% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-spelling.R
\name{GetCorrections}
\alias{GetCorrections}
\title{\code{GetCorrections}}
\usage{
GetCorrections(tokens, counts, spelling.errors, do.not.correct)
}
\arguments{
\item{tokens}{A character vector containing the set of unique tokens identified
from a set of text.}

\item{counts}{A numeric vector containing the number of occurrences of each token
within the text.}

\item{spelling.errors}{A binary vector indicating which of the tokens have been
identified as mis-spelled. Obtained, for example, from \code{\link{FindSpellingErrors}}}

\item{do.not.correct}{A binary vector indicating any tokens which should not be corrected.}
}
\value{
A character vector containing one correction for each original token.
}
\description{
Search for corrections for mispelled words within the set of words
             identified from some text.
}
\details{
This function is fairly conservative, only returning the most
         obvious corrections based on the following heuristics:
1. The correction must begin with the same letter as the misspelling
2. The correction must be within a distance of 1 from the misspelling
   ie one insertion, one deletion, one replacement.
3. When there is more than one potential correction for a word, including the original misspelled
word, the function chooses the most frequent word as the correction. The original misspelled
word is included as brand names will often not appear in the dictionary, but may appear
frequently in the text.
}
\examples{
GetCorrections(tokens = c("hello", "hullo", "hollo", "hollow"),
               counts = c(10, 5, 4, 5),
               spelling.errors = FindSpellingErrors(c("hello", "hullo", "hollo", "hollow")),
               do.not.correct = c(0, 0, 0, 0))
}

