% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phrases.R
\name{GetMostFrequentPhrases}
\alias{GetMostFrequentPhrases}
\title{\code{GetMostFrequentPhrases}}
\usage{
GetMostFrequentPhrases(text, num.words = c(2, 3), min.frequency = 3)
}
\arguments{
\item{text}{A character vector containg the text responses.}

\item{num.words}{A vector of integers specifying the lengths of phrases to be
considered. For example, \code{c(2,3)} indicates that we want to
look at phrases with two or three words.}

\item{min.frequency}{Don't show phrases that appear less often than this value.}
}
\value{
A data frame whose first column contains phrases, and second column contains
        counts of those phrases in \code{text}.
}
\description{
Create a data frame containing the most frequent phrases in some text.
}
\examples{
GetMostFrequentPhrases(ftaFavoriteThings)

}

