% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-stemming.R
\name{GetStemNames}
\alias{GetStemNames}
\title{\code{GetStemNames}}
\usage{
GetStemNames(tokens, counts)
}
\arguments{
\item{tokens}{A character vector containing the tokens to be mapped.}

\item{counts}{A numeric vector representing the counts of each word.}
}
\value{
A character vector with the same length as \code{tokens} that contains the
        new mapped tokens.
}
\description{
Stem a set of tokens and map each token to the most frequent
             token in the set which also has the same stem.
}
\details{
On it's own, stemming can often produce text which is not readable
         because word stems are often not real words.

         This function uses the Snowball stemming algorithm from the package
         \code{SnowballC} and some additional heuristics to map each token
         to the token in the set which has the same stem and the highest count.
         For example, if the tokens \code{search} and \code{searched} have
         counts of 4 and 5 in the text, then both will be mapped to \code{searched}
         as it appears more frequently.
}

