% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qhelperfunctions.R
\name{GetTextAnalysisOptions}
\alias{GetTextAnalysisOptions}
\title{\code{GetTextAnalysisOptions}}
\usage{
GetTextAnalysisOptions(phrases = "", extra.stopwords.text = "",
  replacements.text = "", do.stem = FALSE, do.spell = FALSE)
}
\arguments{
\item{phrases}{A string which contains pairs of words, separated by commas, to be treated as phrases
rather than as separate words. For example: "hello world, very happy" would instruct the options
to treat appearances of "hello world" and "very happy" as single units, rather than as the individual
words "hello", "world", "very", and "happy".}

\item{extra.stopwords.text}{A string containing additional words to be removed from the analysis. The
words should be separated by commas.}

\item{replacements.text}{A string containing instructions on how to replace particular words. The
string should be in the form \code{"<word to replace 1>:<replacement word 1>, <word to replace 2>:<replacement word 2>"}.
So to replace "good" with "great", and "like" with "love", the string should be \code{"good:great,like:love"}.}

\item{do.stem}{A boolean value indicating whether words should be stemmed.}

\item{do.spell}{A boolean value indicating whether words should be spell-checked and corrected.}
}
\description{
This function takes a set of options specified by the user, combines them with the
standard options for text-cleaning, and generates a simple sequence of operations for the creation
of a word bag. Word bag creation can take an arbitrary sequence, and this function interprets
the user's inputs and converts it into a sensible, simple sequence.
}
