% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-mainWrappers.R
\name{MostFrequentWords}
\alias{MostFrequentWords}
\title{\code{MostFrequentWords}}
\usage{
MostFrequentWords(x, min.frequency = 2, alphabetical = NULL)
}
\arguments{
\item{x}{A character vector containing the text that you want to process
or an object of class \code{wordBag} created, for example, by
\code{\link{InitializeWordBag}}.}

\item{min.frequency}{An integer specifying the minium count to keep in the output.}

\item{alphabetical}{A boolean value specifying whether you want the output to be
                    sorted alphabetically or by count.}
}
\value{
A vector of integers, where the name of each element corresponds to the word
        and the value corresponds to the count.
}
\description{
Count the words appearing in text.
}
\examples{
MostFrequentWords(ftaFavoriteThings, min.frequency = 3)
}

