% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nasaTweetText}
\alias{nasaTweetText}
\title{Text from tweets about NASA Announcement - Water on Marse.}
\format{A character vector with 3554 entries.}
\usage{
nasaTweetText
}
\description{
Text from tweets about NASA Announcement - Water on Marse.
}
\keyword{datasets}

