% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{SaveNetSentimentScores}
\alias{SaveNetSentimentScores}
\title{\code{SaveNetSentimentScores}}
\usage{
SaveNetSentimentScores(input)
}
\arguments{
\item{input}{Either a character vector or object of class \code{wordBag}.}
}
\value{
A vector of integers showing the number of positive words minus the number of negative
        words for each text entry.
}
\description{
Return the difference between the number of positive and negative words for
             a character vector or wordBag object.
}
\details{
If the input is a character vector then the text will be tokenized to compute the
         sentiment, but no additional cleaning is done on the text. If the input is a \code{wordBag}
         then the sentiment will be computed based on the transformed text (i.e. using the
         results of the cleaning that has been done by the wordBag).
}

