% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{SaveTidiedText}
\alias{SaveTidiedText}
\title{\code{SaveTidiedText}}
\usage{
SaveTidiedText(word.bag)
}
\arguments{
\item{word.bag}{An object of class \code{wordBag} (created by \code{\link{InitializeWordBag}}.}
}
\value{
A character vector containing the tidied text.
}
\description{
Return the tidied text from a wordBag. This text is the result of the processing
done by the wordBag.
}
