% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sentiment.R
\name{TagSentiment}
\alias{TagSentiment}
\title{\code{TagSentiment}}
\usage{
TagSentiment(tokens, pos.words = ftaPositiveWords,
  neg.words = ftaNegativeWords)
}
\arguments{
\item{tokens}{A character vector containing the tokens to tag.}

\item{pos.words}{A character vector containing the dictionary of positive words.}

\item{neg.words}{A character vector containing the dictionary of negative words.}
}
\value{
A character vector with the same length as \code{tokens} which has entries
         \code{[pos]} for postive words, \code{[neg]} for negative words, and
         \code{[neu]} for neutral words (i.e. not found in either dictionary.)
}
\description{
Identify words as positive, negative, or neutral.
}
\details{
This function does not handle negation, but only looks for each token with
the two dictionaries. Negation can be handled later by looking at the context of each
token in the original text, as done by \code{\link{ScoreSentimentForString}}.
}
\examples{
TagSentiment(c("great","happy","the", "of", "bad", "terrible"))
}

