% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trumpTweets}
\alias{trumpTweets}
\title{Twitter Data Set from Donald Trump as used on
http://varianceexplained.org/r/trump-tweets/}
\format{A data frame with 1512 rows.}
\usage{
trumpTweets
}
\description{
Twitter Data Set from Donald Trump as used on
http://varianceexplained.org/r/trump-tweets/
}
\keyword{datasets}
