% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{AsTermMatrix}
\alias{AsTermMatrix}
\title{\code{AsTermMatrix}}
\usage{
AsTermMatrix(x, min.frequency = 5, remove.stopwords = TRUE,
  stoplist = get("ftaStopList"), operations = c("spelling", "stemming"),
  spelling.dictionary = get("ftaDictionary"), manual.replacements = NULL,
  sparse = FALSE)
}
\arguments{
\item{x}{Either a character vector or a WordBag object, created for example by the function InitializeWordBag.
If \code{x} is a character vector then additional parameters can be supplied for processing the
text before scoring sentiment. For a description of the options see \code{\link{InitializeWordBag}}.}

\item{min.frequency}{An integer specifiying the smallest frequency of word to keep in the transformed text.}

\item{remove.stopwords}{A boolean value specifying whether or not to identify stopwords and remove them from subsequent analyses.}

\item{stoplist}{A character vector containg the stopwords. The default value is this package's built-in stopwords list, ftaStopList.}

\item{operations}{A character vector specifying the sequence of operations to be performed on the text. Valid operations are
\code{"spelling"}, \code{"stemming"}, and \code{"replacement"} (for manually replacing words).}

\item{spelling.dictionary}{A character vector containing the dictionary to use to check each word for mis-spellings. The default value is this package's built-in english dictionary, ftaDictionary.}

\item{manual.replacements}{A matrix of characters with two columms. The first column specifies the words to replace, and the second column specifies the corresponding replacements.}

\item{sparse}{Whether to return the term matrix as a sparse matrix.}
}
\value{
A \code{matrix} with one row for each text response (referred to as a \code{document}) and once column for
         each word or phrase extracted from the text.
}
\description{
Generate a matrix of sentiment scores from a character vector or wordBag object.
}
\examples{
AsTermMatrix(ftaFavoriteThings)

}
