% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftaStopList}
\alias{ftaStopList}
\title{List of English language stopwords.}
\format{A character vector with 128 entries.}
\usage{
ftaStopList
}
\description{
List of English language stopwords.
}
\keyword{datasets}
