% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-wordBag.R
\name{InitializeWordBag}
\alias{InitializeWordBag}
\title{Create a bag-of-words}
\usage{
InitializeWordBag(text, remove.stopwords = TRUE, stoplist = ftaStopList,
  correct.spelling = TRUE, spelling.dictionary = ftaDictionary,
  do.stemming = TRUE)
}
\arguments{
\item{text}{A character vector containing the text to be analyzed.}

\item{remove.stopwords}{A boolean value specifying whether or not to identify stopwords and remove them from subsequent analyses.}

\item{stoplist}{A character vector containg the stopwords. The default value is this package's built-in stopwords list, ftaStopList.}

\item{correct.spelling}{A boolean value specifying whether or not to identify spelling mistakes and to generate corrections.}

\item{spelling.dictionary}{A character vector containing the dictionary to use to check each word for mis-spellings. The default value is this package's built-in english dictionary, ftaDictionary.}

\item{do.stemming}{A boolean value specifying whether or not to stem the words and replace each stem with the most commonly-occuring word that matches that stem.}
}
\description{
Create a bag-of-words object from a vector of strings.
}

