% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flipTextAnalytics-wordBag.R
\name{InitializeWordBag}
\alias{InitializeWordBag}
\title{\code{InitializeWordBag}}
\usage{
InitializeWordBag(text, remove.stopwords = TRUE, stoplist = ftaStopList,
  operations = c("spelling", "stemming"),
  spelling.dictionary = ftaDictionary, manual.replacements = NULL,
  phrases = NULL, min.frequency = 1, alphabetical.sort = TRUE,
  print.type = "frequencies")
}
\arguments{
\item{text}{A character vector containing the text to be analyzed.}

\item{remove.stopwords}{A boolean value specifying whether or not to identify stopwords and remove them from subsequent analyses.}

\item{stoplist}{A character vector containg the stopwords. The default value is this package's built-in stopwords list, ftaStopList.}

\item{spelling.dictionary}{A character vector containing the dictionary to use to check each word for mis-spellings. The default value is this package's built-in english dictionary, ftaDictionary.}

\item{manual.replacements}{A matrix of characters with two columms. The first column specifies the words to replace, and the second column specifies the corresponding replacements.}

\item{phrases}{A character vector containing strings that should be treated as single units during processing.}

\item{min.frequency}{An integer specifiying the smallest frequency of word to keep in the transformed text.}

\item{alphabetical.sort}{A logical value indicating whether word lists associated with this Word Bag should be printed in alphabetical order or in order of word frequencies.}

\item{print.type}{A string indicating the type of printing that should be performed when the Word Bag is printed. The two options are "frequencies", which will print the unique tokens next to their frequencies, and "transformations", which will print the original text next to the new transformed text. All printing is done using datatable from package D3.}

\item{correct.spelling}{A boolean value specifying whether or not to identify spelling mistakes and to generate corrections.}

\item{do.stemming}{A boolean value specifying whether or not to stem the words and replace each stem with the most commonly-occuring word that matches that stem.}
}
\value{
An object of class \code{wordBag} containing the word bag details.
}
\description{
Create a wordBag object from a vector of strings.
}

