% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{SaveNetSentimentScores}
\alias{SaveNetSentimentScores}
\title{\code{SaveNetSentimentScores}}
\usage{
SaveNetSentimentScores(input, check.simple.suffixes = FALSE,
  simple.suffixes = c("s", "es", "ed", "d", "ing"),
  pos.words = ftaPositiveWords, neg.words = ftaNegativeWords)
}
\arguments{
\item{input}{Either a character vector or object of class \code{wordBag}.}

\item{check.simple.suffixes}{If true, whevenver a token is not found in
pos.words AND neg.words, the algorithm will attempt to trim suffixes
from the word and check the dictionaries again. This is to account for
cases where the dictionaries do not contain variants of sentiment words
(plurals, past tenses, etc). The suffixes to check are specified in the
next argument.}

\item{simple.suffixes}{An array of strings containing suffixes to strip when checking
for words among the positive and negative word lists.}

\item{pos.words}{A character vector containing the dictionary of positive words.}

\item{neg.words}{A character vector containing the dictionary of negative words.}
}
\value{
A vector of integers showing the number of positive words minus the number of negative
        words for each text entry.
}
\description{
Return the difference between the number of positive and negative words for
             a character vector or wordBag object.
}
\details{
If the input is a character vector then the text will be tokenized to compute the
         sentiment, but no additional cleaning is done on the text. If the input is a \code{wordBag}
         then the sentiment will be computed based on the transformed text (i.e. using the
         results of the cleaning that has been done by the wordBag).
}

